Make files for each class and separate them
한 파일에 클래스가 많아 수정이 어려워 클래스별로 새로운 파일을 만들어 분리
딜러와 플레이어는 둘다 카드들을 가지고 있다는 특성이 공통되어 부모 클래스를 만들어 관리를 효율적으로 하고자 했음
