+++ week1-poker/PokerApp/Player.swift
@@ -1,0 +1,17 @@
+//
+//  Player.swift
+//  PokerApp
+//
+//  Created by 윤준수 on 2021/01/07.
+//
+
+import Foundation
+class Player: Person {
+    private let id: Int
+    
+    init(id: Int) {
+        self.id = id
+    }
+    
+    
+}
